backend/requirements.txt
# ============================================================
# KNOT Backend — Python Dependencies
# ============================================================

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.2
geoalchemy2==0.15.1
psycopg2-binary==2.9.9

# Validation
pydantic==2.8.2
pydantic-settings==2.3.4
email-validator==2.1.1
bleach==6.1.0
python-magic-bin==0.4.14   # MIME type validation for image uploads (Windows bundled)

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis
redis==5.0.7

# Object Storage
boto3==1.34.144
minio==7.2.7

# Email
aiosmtplib==3.0.1

# Rate limiting
slowapi==0.1.9

# Retry logic
tenacity==8.3.0

# HTTP client
httpx==0.27.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
